Check number of args.

+++ src/redlot/t_kv.go
@@ -1,6 +1,9 @@
 package redlot
 
-import "fmt"
+import (
+	"errors"
+	"fmt"
+)
 
 func encode_kv_key(key []byte) (buf []byte) {
 	buf = append(buf, TYPE_KV)
@@ -17,20 +20,36 @@ func decode_kv_key(buf []byte) (key []byte) {
 }
 
 func Get(args [][]byte) (interface{}, error) {
+	if len(args) < 1 {
+		return "", ERR_NOS_ARGS
+	}
+
 	v, err := db.Get(encode_kv_key(args[0]), nil)
 	return string(v), err
 }
 
 func Set(args [][]byte) (interface{}, error) {
+	if len(args) < 2 {
+		return nil, ERR_NOS_ARGS
+	}
+
 	fmt.Printf("SET %s %s\n", args[0], args[1])
 	return nil, db.Put(encode_kv_key(args[0]), args[1], nil)
 }
 
 func Del(args [][]byte) (interface{}, error) {
+	if len(args) < 1 {
+		return nil, ERR_NOS_ARGS
+	}
+
 	return nil, db.Delete(encode_kv_key(args[0]), nil)
 }
 
 func Exists(args [][]byte) (interface{}, error) {
+	if len(args) < 1 {
+		return int64(-1), ERR_NOS_ARGS
+	}
+
 	ret, err := db.Has(encode_kv_key(args[0]), nil)
 	if ret {
 		return int64(1), err
